give feedback when metamask not found
