+++ apps/web/src/useLastAddress.tsx
@@ -1,0 +1,12 @@
+import { useEffect, useState } from 'react'
+import { getAddresses } from 'src/history'
+
+export function useLastAddress() {
+  const [lastAddress, setLastAddress] = useState<string>()
+  useEffect(() => {
+    const lastUsedAddress = getAddresses().at(0)
+    setLastAddress(lastUsedAddress)
+  }, [])
+
+  return lastAddress
+}
